Can't trust the relative time nhentai gives in its HTML
Whatever server-side logic generates the relative upload date shown on
nhentai galleries is not great at rounding. My test gallery was uploaded
on June 8, and I'm testing on August 9, but the HTML still says "1 month
ago" instead of "2 months ago" (which JavaScript updates the page to say
when loaded in a browser).

So I guess this plugin has to do its own work in order for the relative
time displayed to be reasonable and not make users go "????".